Move model stubs into stubs package
