+++ src/test/java/ay2122s1_cs2103t_w16_2/btbb/testutil/stubs/ModelStub.java
@@ -1,0 +1,102 @@
+package ay2122s1_cs2103t_w16_2.btbb.testutil.stubs;
+
+import java.nio.file.Path;
+import java.util.function.Predicate;
+
+import ay2122s1_cs2103t_w16_2.btbb.commons.core.GuiSettings;
+import ay2122s1_cs2103t_w16_2.btbb.model.Model;
+import ay2122s1_cs2103t_w16_2.btbb.model.ReadOnlyAddressBook;
+import ay2122s1_cs2103t_w16_2.btbb.model.ReadOnlyUserPrefs;
+import ay2122s1_cs2103t_w16_2.btbb.model.client.Client;
+import ay2122s1_cs2103t_w16_2.btbb.model.order.Order;
+import javafx.collections.ObservableList;
+
+/**
+ * A default model stub that has all of its methods failing.
+ */
+public class ModelStub implements Model {
+    @Override
+    public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ReadOnlyUserPrefs getUserPrefs() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public GuiSettings getGuiSettings() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setGuiSettings(GuiSettings guiSettings) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public Path getAddressBookFilePath() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setAddressBookFilePath(Path addressBookFilePath) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setAddressBook(ReadOnlyAddressBook newData) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ReadOnlyAddressBook getAddressBook() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void updateFilteredOrderList(Predicate<Order> predicate) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addClient(Client client) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasClient(Client client) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void deleteClient(Client target) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void setClient(Client target, Client editedClient) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public ObservableList<Client> getFilteredClientList() {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void updateFilteredClientList(Predicate<Client> predicate) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public void addOrder(Order order) {
+        throw new AssertionError("This method should not be called.");
+    }
+
+    @Override
+    public boolean hasOrder(Order order) {
+        throw new AssertionError("This method should not be called.");
+    }
+}
